refactor: screen director
