--- lib/features/characters/list/character_list_screen.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-class CharacterListScreen extends ConsumerStatefulWidget {
-  const CharacterListScreen({super.key});
-
-  @override
-  ConsumerState<CharacterListScreen> createState() =>
-      _CharacterListScreenState();
-}
-
-class _CharacterListScreenState extends ConsumerState<CharacterListScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Container();
-  }
-}
